Replaced std::vector<uint8_t> with Data in many files and adjusted referrer codes. Added type casts in Trinity.cpp, a2r.cpp, hfe.cpp, mfi.cpp, scl.cpp. Commented unused deviceReadingPolicy parameter. Fixed destructors by adding the override keyword where it was required. Removed obsolete includes.

+++ src/FluxDecoder.cpp
@@ -121,5 +121,5 @@ int FluxDecoder::next_flux()
     }
 
     auto time_ns = *m_flux_it++;
-    return time_ns;
+    return static_cast<int>(time_ns);
 }
